requirements.txt
pyyaml
jsonschema>=4.18.0
pytest
